Move circular reference tests into separate file
