+++ tests/Unit/FactoryTest.php
@@ -12,7 +12,6 @@ use Yiisoft\Definitions\ArrayDefinition;
 use Yiisoft\Definitions\DynamicReference;
 use Yiisoft\Definitions\Reference;
 use Yiisoft\Definitions\ValueDefinition;
-use Yiisoft\Definitions\Exception\CircularReferenceException;
 use Yiisoft\Definitions\Exception\InvalidConfigException;
 use Yiisoft\Definitions\Exception\NotInstantiableClassException;
 use Yiisoft\Definitions\Exception\NotInstantiableException;
@@ -21,12 +20,7 @@ use Yiisoft\Factory\NotFoundException;
 use Yiisoft\Factory\Tests\Support\CallableDependency;
 use Yiisoft\Factory\Tests\Support\Car;
 use Yiisoft\Factory\Tests\Support\CarFactory;
-use Yiisoft\Factory\Tests\Support\Circular\Chicken;
-use Yiisoft\Factory\Tests\Support\Circular\CircularA;
-use Yiisoft\Factory\Tests\Support\Circular\CircularB;
-use Yiisoft\Factory\Tests\Support\Circular\TreeItem;
 use Yiisoft\Factory\Tests\Support\ColorInterface;
-use Yiisoft\Factory\Tests\Support\Circular\Egg;
 use Yiisoft\Factory\Tests\Support\ExcessiveConstructorParameters;
 use Yiisoft\Factory\Tests\Support\Firefighter;
 use Yiisoft\Factory\Tests\Support\ColorPink;
@@ -1398,120 +1392,6 @@ final class FactoryTest extends TestCase
 
         $this->expectException(InvalidConfigException::class);
         $factory->create('test');
-    }
-
-    public function testOptionalCircularClassDependency(): void
-    {
-        $factory = new Factory(new SimpleContainer(), [
-            CircularA::class => CircularA::class,
-            CircularB::class => CircularB::class,
-        ]);
-
-        $a = $factory->create(CircularA::class);
-
-        $this->assertInstanceOf(CircularB::class, $a->b);
-        $this->assertNull($a->b->a);
-    }
-
-    public function testCircularClassDependency(): void
-    {
-        $factory = new Factory(
-            new SimpleContainer(),
-            [
-                Chicken::class => Chicken::class,
-                Egg::class => Egg::class,
-            ]
-        );
-
-        $this->expectException(CircularReferenceException::class);
-        $factory->create(Chicken::class);
-    }
-
-    public function testCircularClassDependencyWithoutDefinition(): void
-    {
-        $factory = new Factory(new SimpleContainer(), [
-            Egg::class => Egg::class,
-            Chicken::class => Chicken::class,
-        ]);
-
-        $this->expectException(CircularReferenceException::class);
-        $factory->create(Chicken::class);
-    }
-
-    public function testCircularReferences(): void
-    {
-        $factory = new Factory(
-            new SimpleContainer(),
-            [
-                'engine-1' => Reference::to('engine-2'),
-                'engine-2' => Reference::to('engine-3'),
-                'engine-3' => Reference::to('engine-1'),
-            ]
-        );
-
-        $this->expectException(CircularReferenceException::class);
-        $this->expectExceptionMessage(
-            'Circular reference to "engine-2" detected while creating: engine-2, engine-3, engine-1.'
-        );
-        $factory->create('engine-1');
-    }
-
-    public function testCircularReference(): void
-    {
-        $factory = new Factory(new SimpleContainer(), [TreeItem::class => TreeItem::class]);
-
-        $this->expectException(CircularReferenceException::class);
-        $factory->create(TreeItem::class);
-    }
-
-    /**
-     * @link https://github.com/yiisoft/di/pull/189
-     */
-    public function testFalsePositiveCircularReferenceWithClassID(): void
-    {
-        $this->expectNotToPerformAssertions();
-
-        $factory = new Factory(new SimpleContainer());
-
-        // Build an object
-        $factory->create(ColorPink::class);
-
-        // set definition to container
-        (fn (string $id, $definition) => $this->set($id, $definition))->call(
-            $factory,
-            ColorPink::class,
-            ColorPink::class
-        );
-
-        try {
-            $factory->create(ColorPink::class);
-        } catch (CircularReferenceException $e) {
-            $this->fail('Circular reference detected false positively.');
-        }
-    }
-
-    /**
-     * @link https://github.com/yiisoft/di/pull/189
-     */
-    public function testFalsePositiveCircularReferenceWithStringID(): void
-    {
-        $this->expectNotToPerformAssertions();
-
-        $factory = new Factory(new SimpleContainer());
-        try {
-            $factory->create('test');
-        } catch (NotFoundException $e) {
-            // It is expected
-        }
-
-        // set definition to container
-        (fn (string $id, $definition) => $this->set($id, $definition))->call($factory, 'test', ColorPink::class);
-
-        try {
-            $factory->create('test');
-        } catch (CircularReferenceException $e) {
-            $this->fail('Circular reference detected false positively.');
-        }
     }
 
     public function testNullableScalarConstructorArgument(): void
